+++ packages/avilatek_core/lib/common/core.dart
@@ -1,5 +1,0 @@
-/// Core utilities commonly used in Avila Tek projects.
-library core;
-
-export 'assertions.dart';
-export 'error_radar.dart';