+++ packages/avilatek_core/lib/src/interfaces/device_storage.dart
@@ -4,7 +4,7 @@
 /// device storage, like
 /// [shared_preferences](https://pub.dev/packages/shared_preferences) or
 /// [flutter_secure_storage](https://pub.dev/packages/flutter_secure_storage).
-abstract class IDeviceStorage {
+abstract class DeviceStorage {
   /// Stores the provided [value] under the given [key] name.
   Future<void> setValue({
     required String key,